requirements.txt
# API Clients
requests>=2.31.0
feedparser>=6.0.10
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.4.0

# PDF Processing
PyMuPDF>=1.23.0
weasyprint>=60.0

# Data Processing
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
rich>=13.7.0
tenacity>=8.2.0
click>=8.1.0

# XML Parsing (for PubMed)
lxml>=5.1.0
